+++ docs/reference/changelog.rst
@@ -33,6 +33,10 @@ Bug fixes:
   ``<distributionManagement>`` tags.
 - With the maven-use plugin, fix Maven plugin versioning errors when indirectly depending
   on a plugin.
+- Fix content migration when using :class:`~craft_parts.FilesystemMount` during the
+  prime step. Instead of relying on the content of the overlay, gather the list of
+  files and directories from the stage directory and the state of the stage
+  migration.
 
 .. _release-2.14.0: